Abort update-changelog workflow if target branch does not match major release version (#13)
* Add guard to update-changelog Workflow

These new steps validate if the given release version (eg. `v10.1.2`) matches with the selected target branch (eg. `10.x`).

The script extracts the major version from both strings by using parameter expansion in Bash.
The `v` is being removed from the given release version (resulting in `10.1.2`). Next the major version is extracted by ignoring all values after the first dot (resulting in `10`).

The major version is also extracted from the selected branch name by ignoring all values after the first dot (resulting in `10`).

If the two major version do not match the second step is being executed, a `VERSION_MISMATCH` output variable is set to true.

If that's the case the next step is executed.
By using the `actions/github-script`-Action we mark the Workflow run as failed.[1].
The GitHub user triggering the workflow should receive a notification by mail and see the error message on the summary screen of the Workflow run.

[1]: https://github.com/actions/toolkit/tree/main/packages/core#exit-codes

* Don't run guard when main/master is targeted